+++ Front/TemplatesForm.html
@@ -271,7 +271,7 @@
   <!-- END INBOX -->
 </div>
 </div>
-<script src="http://localhost:3000/js/templatesForm.js">
+<script type="module" src="http://localhost:3000/js/templatesForm.js">
 
 </script>
 </body>
