+++ Front/homepage.html
@@ -241,12 +241,7 @@
         <!-- END INBOX -->
         </div>
     </div>
-    <script>
-        clickSend();
-        getAllTemplates();
-        </script>
-      </div>
-  </body>
+</body>
 </html>
