+++ Front/schedule.html
@@ -11,6 +11,7 @@
     <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" rel="stylesheet">
     <script type="module" src="http://localhost:3000/js/APIpaths.js"></script>
     <script type="module" src="http://localhost:3000/js/scripts.js"></script>
+    <script type="module" src="http://localhost:3000/js/schedScript.js"></script>
 
 </head>
 <body>
@@ -72,10 +73,6 @@
                             <div class="padding"></div>
                             <div class="table-responsive">
                                 <section id="chart" class="table">
-                                    <script>
-                                        getScheduledEmails();
-                                    </script>
-
                                 </section>
                             </div>
                             <ul class="pagination">
@@ -244,10 +241,6 @@
         </div>
     </div>
 </div>
-<script>
-    getTotaltoSendEmail();
-    createTable();
-</script>
 </body>
 </html>
