+++ Front/sent.html
@@ -11,6 +11,7 @@
     <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" rel="stylesheet">
     <script type="module" src="http://localhost:3000/js/APIpaths.js"></script>
     <script type="module" src="http://localhost:3000/js/scripts.js"></script>
+    <script type="module" src="http://localhost:3000/js/sentScript.js"></script>
 </head>
 <body>
 <div class="container">
@@ -234,13 +235,6 @@
     </div>
     <!-- END INBOX -->
 </div>
-    </div>
-
-<script>
-    getSentEmail();
-    getTotalSentEmail();
-    clickSend();
-</script>
 </body>
 
 </html>
